uart now uses size_t

+++ include/uart.h
@@ -11,6 +11,6 @@ void uart_puts(const uint8_t* str);
 void uart_printf(const uint8_t *format, ...);
 
 int8_t uart_getc();
-void uart_gets(uint8_t* buf, uint32_t len);
+void uart_gets(uint8_t* buf, size_t len);
 
 #endif /*UART_H*/
